+++ replacer-frontend/src/app/api/models.ts
@@ -1,3 +1,5 @@
+/* tslint:disable */
+/* eslint-disable */
 export { ReviewPage } from './models/review-page';
 export { ReviewSection } from './models/review-section';
 export { ReviewedReplacement } from './models/reviewed-replacement';