+++ replacer-frontend/src/app/api/models/kind-count.ts
@@ -10,10 +10,10 @@ export interface KindCount {
   /**
    * Replacement kind code
    */
-  'k': number;
+  k: number;
 
   /**
    * Count of pages to review grouped by subtype for a given kind
    */
-  'l': Array<SubtypeCount>;
+  l: Array<SubtypeCount>;
 }