+++ replacer-frontend/src/app/api/services/administration-api.service.ts
@@ -1,30 +1,24 @@
 /* tslint:disable */
 /* eslint-disable */
+import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse, HttpContext } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+
 import { BaseService } from '../base-service';
 import { ApiConfiguration } from '../api-configuration';
 import { StrictHttpResponse } from '../strict-http-response';
 import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
 
 import { PublicIp } from '../models/public-ip';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable({ providedIn: 'root' })
 export class AdministrationApiService extends BaseService {
-  constructor(
-    config: ApiConfiguration,
-    http: HttpClient
-  ) {
+  constructor(config: ApiConfiguration, http: HttpClient) {
     super(config, http);
   }
 
-  /**
-   * Path part for operation getPublicIp
-   */
+  /** Path part for operation `getPublicIp()` */
   static readonly GetPublicIpPath = '/api/admin/public-ip';
 
   /**
@@ -37,21 +31,19 @@ export class AdministrationApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  getPublicIp$Response(params?: {
+  getPublicIp$Response(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<PublicIp>> {
-
+  ): Observable<StrictHttpResponse<PublicIp>> {
     const rb = new RequestBuilder(this.rootUrl, AdministrationApiService.GetPublicIpPath, 'get');
     if (params) {
     }
 
-    return this.http.request(rb.build({
-      responseType: 'json',
-      accept: 'application/json',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'json', accept: 'application/json', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<PublicIp>;
       })
@@ -63,18 +55,18 @@ export class AdministrationApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `getPublicIp$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  getPublicIp(params?: {
+  getPublicIp(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<PublicIp> {
-
-    return this.getPublicIp$Response(params).pipe(
-      map((r: StrictHttpResponse<PublicIp>) => r.body as PublicIp)
+  ): Observable<PublicIp> {
+    return this.getPublicIp$Response(params, context).pipe(
+      map((r: StrictHttpResponse<PublicIp>): PublicIp => r.body)
     );
   }