+++ replacer-frontend/src/app/api/services/authentication-api.service.ts
@@ -1,32 +1,26 @@
 /* tslint:disable */
 /* eslint-disable */
+import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse, HttpContext } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+
 import { BaseService } from '../base-service';
 import { ApiConfiguration } from '../api-configuration';
 import { StrictHttpResponse } from '../strict-http-response';
 import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
 
 import { InitiateAuthenticationResponse } from '../models/initiate-authentication-response';
 import { User } from '../models/user';
 import { VerifyAuthenticationRequest } from '../models/verify-authentication-request';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable({ providedIn: 'root' })
 export class AuthenticationApiService extends BaseService {
-  constructor(
-    config: ApiConfiguration,
-    http: HttpClient
-  ) {
+  constructor(config: ApiConfiguration, http: HttpClient) {
     super(config, http);
   }
 
-  /**
-   * Path part for operation verifyAuthentication
-   */
+  /** Path part for operation `verifyAuthentication()` */
   static readonly VerifyAuthenticationPath = '/api/authentication/verify';
 
   /**
@@ -39,23 +33,21 @@ export class AuthenticationApiService extends BaseService {
    *
    * This method sends `application/json` and handles request body of type `application/json`.
    */
-  verifyAuthentication$Response(params: {
+  verifyAuthentication$Response(
+    params: {
+      body: VerifyAuthenticationRequest
+    },
     context?: HttpContext
-    body: VerifyAuthenticationRequest
-  }
-): Observable<StrictHttpResponse<User>> {
-
+  ): Observable<StrictHttpResponse<User>> {
     const rb = new RequestBuilder(this.rootUrl, AuthenticationApiService.VerifyAuthenticationPath, 'post');
     if (params) {
       rb.body(params.body, 'application/json');
     }
 
-    return this.http.request(rb.build({
-      responseType: 'json',
-      accept: 'application/json',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'json', accept: 'application/json', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<User>;
       })
@@ -67,25 +59,23 @@ export class AuthenticationApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `verifyAuthentication$Response()` instead.
    *
    * This method sends `application/json` and handles request body of type `application/json`.
    */
-  verifyAuthentication(params: {
+  verifyAuthentication(
+    params: {
+      body: VerifyAuthenticationRequest
+    },
     context?: HttpContext
-    body: VerifyAuthenticationRequest
-  }
-): Observable<User> {
-
-    return this.verifyAuthentication$Response(params).pipe(
-      map((r: StrictHttpResponse<User>) => r.body as User)
+  ): Observable<User> {
+    return this.verifyAuthentication$Response(params, context).pipe(
+      map((r: StrictHttpResponse<User>): User => r.body)
     );
   }
 
-  /**
-   * Path part for operation initiateAuthentication
-   */
+  /** Path part for operation `initiateAuthentication()` */
   static readonly InitiateAuthenticationPath = '/api/authentication/initiate';
 
   /**
@@ -98,21 +88,19 @@ export class AuthenticationApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  initiateAuthentication$Response(params?: {
+  initiateAuthentication$Response(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<InitiateAuthenticationResponse>> {
-
+  ): Observable<StrictHttpResponse<InitiateAuthenticationResponse>> {
     const rb = new RequestBuilder(this.rootUrl, AuthenticationApiService.InitiateAuthenticationPath, 'get');
     if (params) {
     }
 
-    return this.http.request(rb.build({
-      responseType: 'json',
-      accept: 'application/json',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'json', accept: 'application/json', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<InitiateAuthenticationResponse>;
       })
@@ -124,18 +112,18 @@ export class AuthenticationApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `initiateAuthentication$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  initiateAuthentication(params?: {
+  initiateAuthentication(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<InitiateAuthenticationResponse> {
-
-    return this.initiateAuthentication$Response(params).pipe(
-      map((r: StrictHttpResponse<InitiateAuthenticationResponse>) => r.body as InitiateAuthenticationResponse)
+  ): Observable<InitiateAuthenticationResponse> {
+    return this.initiateAuthentication$Response(params, context).pipe(
+      map((r: StrictHttpResponse<InitiateAuthenticationResponse>): InitiateAuthenticationResponse => r.body)
     );
   }