+++ replacer-frontend/src/app/api/services/dump-indexing-api.service.ts
@@ -1,30 +1,24 @@
 /* tslint:disable */
 /* eslint-disable */
+import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse, HttpContext } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+
 import { BaseService } from '../base-service';
 import { ApiConfiguration } from '../api-configuration';
 import { StrictHttpResponse } from '../strict-http-response';
 import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
 
 import { DumpIndexingStatus } from '../models/dump-indexing-status';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable({ providedIn: 'root' })
 export class DumpIndexingApiService extends BaseService {
-  constructor(
-    config: ApiConfiguration,
-    http: HttpClient
-  ) {
+  constructor(config: ApiConfiguration, http: HttpClient) {
     super(config, http);
   }
 
-  /**
-   * Path part for operation getDumpIndexingStatus
-   */
+  /** Path part for operation `getDumpIndexingStatus()` */
   static readonly GetDumpIndexingStatusPath = '/api/dump-indexing';
 
   /**
@@ -37,21 +31,19 @@ export class DumpIndexingApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  getDumpIndexingStatus$Response(params?: {
+  getDumpIndexingStatus$Response(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<DumpIndexingStatus>> {
-
+  ): Observable<StrictHttpResponse<DumpIndexingStatus>> {
     const rb = new RequestBuilder(this.rootUrl, DumpIndexingApiService.GetDumpIndexingStatusPath, 'get');
     if (params) {
     }
 
-    return this.http.request(rb.build({
-      responseType: 'json',
-      accept: 'application/json',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'json', accept: 'application/json', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<DumpIndexingStatus>;
       })
@@ -63,24 +55,22 @@ export class DumpIndexingApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `getDumpIndexingStatus$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  getDumpIndexingStatus(params?: {
+  getDumpIndexingStatus(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<DumpIndexingStatus> {
-
-    return this.getDumpIndexingStatus$Response(params).pipe(
-      map((r: StrictHttpResponse<DumpIndexingStatus>) => r.body as DumpIndexingStatus)
+  ): Observable<DumpIndexingStatus> {
+    return this.getDumpIndexingStatus$Response(params, context).pipe(
+      map((r: StrictHttpResponse<DumpIndexingStatus>): DumpIndexingStatus => r.body)
     );
   }
 
-  /**
-   * Path part for operation manualStartDumpIndexing
-   */
+  /** Path part for operation `manualStartDumpIndexing()` */
   static readonly ManualStartDumpIndexingPath = '/api/dump-indexing';
 
   /**
@@ -93,21 +83,19 @@ export class DumpIndexingApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  manualStartDumpIndexing$Response(params?: {
+  manualStartDumpIndexing$Response(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<void>> {
-
+  ): Observable<StrictHttpResponse<void>> {
     const rb = new RequestBuilder(this.rootUrl, DumpIndexingApiService.ManualStartDumpIndexingPath, 'post');
     if (params) {
     }
 
-    return this.http.request(rb.build({
-      responseType: 'text',
-      accept: '*/*',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'text', accept: '*/*', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return (r as HttpResponse<any>).clone({ body: undefined }) as StrictHttpResponse<void>;
       })
@@ -119,18 +107,18 @@ export class DumpIndexingApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `manualStartDumpIndexing$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  manualStartDumpIndexing(params?: {
+  manualStartDumpIndexing(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<void> {
-
-    return this.manualStartDumpIndexing$Response(params).pipe(
-      map((r: StrictHttpResponse<void>) => r.body as void)
+  ): Observable<void> {
+    return this.manualStartDumpIndexing$Response(params, context).pipe(
+      map((r: StrictHttpResponse<void>): void => r.body)
     );
   }