+++ replacer-frontend/src/app/api/services/pages-api.service.ts
@@ -1,30 +1,24 @@
 /* tslint:disable */
 /* eslint-disable */
+import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse, HttpContext } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+
 import { BaseService } from '../base-service';
 import { ApiConfiguration } from '../api-configuration';
 import { StrictHttpResponse } from '../strict-http-response';
 import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
 
 import { KindCount } from '../models/kind-count';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable({ providedIn: 'root' })
 export class PagesApiService extends BaseService {
-  constructor(
-    config: ApiConfiguration,
-    http: HttpClient
-  ) {
+  constructor(config: ApiConfiguration, http: HttpClient) {
     super(config, http);
   }
 
-  /**
-   * Path part for operation reviewPagesByType
-   */
+  /** Path part for operation `reviewPagesByType()` */
   static readonly ReviewPagesByTypePath = '/api/page/type/review';
 
   /**
@@ -37,33 +31,31 @@ export class PagesApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  reviewPagesByType$Response(params: {
+  reviewPagesByType$Response(
+    params: {
 
     /**
      * Replacement kind code
      */
-    kind: number;
+      kind: number;
 
     /**
      * Replacement subtype
      */
-    subtype: string;
+      subtype: string;
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<void>> {
-
+  ): Observable<StrictHttpResponse<void>> {
     const rb = new RequestBuilder(this.rootUrl, PagesApiService.ReviewPagesByTypePath, 'post');
     if (params) {
       rb.query('kind', params.kind, {});
       rb.query('subtype', params.subtype, {});
     }
 
-    return this.http.request(rb.build({
-      responseType: 'text',
-      accept: '*/*',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'text', accept: '*/*', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return (r as HttpResponse<any>).clone({ body: undefined }) as StrictHttpResponse<void>;
       })
@@ -75,34 +67,32 @@ export class PagesApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `reviewPagesByType$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  reviewPagesByType(params: {
+  reviewPagesByType(
+    params: {
 
     /**
      * Replacement kind code
      */
-    kind: number;
+      kind: number;
 
     /**
      * Replacement subtype
      */
-    subtype: string;
+      subtype: string;
+    },
     context?: HttpContext
-  }
-): Observable<void> {
-
-    return this.reviewPagesByType$Response(params).pipe(
-      map((r: StrictHttpResponse<void>) => r.body as void)
+  ): Observable<void> {
+    return this.reviewPagesByType$Response(params, context).pipe(
+      map((r: StrictHttpResponse<void>): void => r.body)
     );
   }
 
-  /**
-   * Path part for operation findPagesToReviewByType
-   */
+  /** Path part for operation `findPagesToReviewByType()` */
   static readonly FindPagesToReviewByTypePath = '/api/page/type';
 
   /**
@@ -115,33 +105,31 @@ export class PagesApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  findPagesToReviewByType$Response(params: {
+  findPagesToReviewByType$Response(
+    params: {
 
     /**
      * Replacement kind code
      */
-    kind: number;
+      kind: number;
 
     /**
      * Replacement subtype
      */
-    subtype: string;
+      subtype: string;
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<string>> {
-
+  ): Observable<StrictHttpResponse<string>> {
     const rb = new RequestBuilder(this.rootUrl, PagesApiService.FindPagesToReviewByTypePath, 'get');
     if (params) {
       rb.query('kind', params.kind, {});
       rb.query('subtype', params.subtype, {});
     }
 
-    return this.http.request(rb.build({
-      responseType: 'text',
-      accept: 'text/plain',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'text', accept: 'text/plain', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<string>;
       })
@@ -153,34 +141,32 @@ export class PagesApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `findPagesToReviewByType$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  findPagesToReviewByType(params: {
+  findPagesToReviewByType(
+    params: {
 
     /**
      * Replacement kind code
      */
-    kind: number;
+      kind: number;
 
     /**
      * Replacement subtype
      */
-    subtype: string;
+      subtype: string;
+    },
     context?: HttpContext
-  }
-): Observable<string> {
-
-    return this.findPagesToReviewByType$Response(params).pipe(
-      map((r: StrictHttpResponse<string>) => r.body as string)
+  ): Observable<string> {
+    return this.findPagesToReviewByType$Response(params, context).pipe(
+      map((r: StrictHttpResponse<string>): string => r.body)
     );
   }
 
-  /**
-   * Path part for operation countPagesNotReviewedByType
-   */
+  /** Path part for operation `countPagesNotReviewedByType()` */
   static readonly CountPagesNotReviewedByTypePath = '/api/page/type/count';
 
   /**
@@ -193,21 +179,19 @@ export class PagesApiService extends BaseService {
    *
    * This method doesn't expect any request body.
    */
-  countPagesNotReviewedByType$Response(params?: {
+  countPagesNotReviewedByType$Response(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<StrictHttpResponse<Array<KindCount>>> {
-
+  ): Observable<StrictHttpResponse<Array<KindCount>>> {
     const rb = new RequestBuilder(this.rootUrl, PagesApiService.CountPagesNotReviewedByTypePath, 'get');
     if (params) {
     }
 
-    return this.http.request(rb.build({
-      responseType: 'json',
-      accept: 'application/json',
-      context: params?.context
-    })).pipe(
-      filter((r: any) => r instanceof HttpResponse),
+    return this.http.request(
+      rb.build({ responseType: 'json', accept: 'application/json', context })
+    ).pipe(
+      filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
       map((r: HttpResponse<any>) => {
         return r as StrictHttpResponse<Array<KindCount>>;
       })
@@ -219,18 +203,18 @@ export class PagesApiService extends BaseService {
    *
    *
    *
-   * This method provides access to only to the response body.
+   * This method provides access only to the response body.
    * To access the full response (for headers, for example), `countPagesNotReviewedByType$Response()` instead.
    *
    * This method doesn't expect any request body.
    */
-  countPagesNotReviewedByType(params?: {
+  countPagesNotReviewedByType(
+    params?: {
+    },
     context?: HttpContext
-  }
-): Observable<Array<KindCount>> {
-
-    return this.countPagesNotReviewedByType$Response(params).pipe(
-      map((r: StrictHttpResponse<Array<KindCount>>) => r.body as Array<KindCount>)
+  ): Observable<Array<KindCount>> {
+    return this.countPagesNotReviewedByType$Response(params, context).pipe(
+      map((r: StrictHttpResponse<Array<KindCount>>): Array<KindCount> => r.body)
     );
   }