+++ replacer-frontend/src/app/app.component.ts
@@ -1,12 +1,13 @@
 import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
 import { HeaderComponent } from './core/header/header.component';
 import { AlertContainerComponent } from './shared/alert/alert-container.component';
 
 @Component({
-  standalone: true,
   selector: 'app-root',
-  imports: [HeaderComponent, AlertContainerComponent, RouterOutlet],
+  standalone: true,
+  imports: [HeaderComponent, AlertContainerComponent, CommonModule, RouterOutlet],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })