+++ replacer-frontend/src/app/app.config.ts
@@ -1,0 +1,26 @@
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { ApplicationConfig, importProvidersFrom } from '@angular/core';
+import { provideRouter } from '@angular/router';
+import { ApiModule } from './api/api.module';
+import { routes } from './app.routes';
+import { AuthInterceptor } from './core/interceptor/auth-interceptor';
+import { LangInterceptor } from './core/interceptor/lang-interceptor';
+import { environment } from '../environments/environment';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideRouter(routes),
+    importProvidersFrom(ApiModule.forRoot({ rootUrl: environment.baseUrl })),
+    provideHttpClient(withInterceptorsFromDi()),
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: LangInterceptor,
+      multi: true
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: AuthInterceptor,
+      multi: true
+    }
+  ]
+};