+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 2.23.7 (??)
+- Fix: intercept any 401 error in a response
+- Upgrade frontend to Angular 16
+
 ## 2.23.6 (8 May 2023)
 - Fix regression: dropdown to change language
 - Fix: don't stop the whole indexing process if there is any error on a specific dump page